backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
supabase==2.13.0
openai==1.59.5
anthropic==0.42.0
python-multipart==0.0.20
